begynt ordentlig refactor av toppmeny og søkefelt for å unngå alle bugs vi har
