--- src/LookupControl/LookupControl.css
@@ -1,12 +1,0 @@
-.toolbar {
-  padding: 0;
-}
-.squareBottom {
-  border-radius: "4px 4px 0 0";
-}
-.darkButton {
-  color: "#616161";
-}
-.lightButton {
-  color: "#b4b4b4";
-}
