+++ src/Searchbar/SearchField/SearchField.js
@@ -1,19 +1,12 @@
-import { Input } from "@material-ui/core";
 import React, { useEffect, useRef } from "react";
 
-const SearchBox = ({
-  isSearching,
-  onBlur,
-  onQueryChange,
-  query,
-  className
-}) => {
+const SearchBox = ({ isSearching, onBlur, onQueryChange, query }) => {
   const inputEl = useRef(null);
   useEffect(() => {
     if (isSearching) inputEl.current.focus();
   }, [isSearching]);
   return (
-    <Input
+    <input
       inputRef={inputEl}
       value={query}
       placeholder={"Søk i Natur i Norge"}
@@ -21,8 +14,6 @@ const SearchBox = ({
       onChange={onQueryChange}
       fullWidth={true}
       disableUnderline={true}
-      classes={{ input: className }}
-      className={className}
     />
   );
 };
